+++ ThirdPartyAdapters/ironsource/ironsource/src/main/java/com/google/ads/mediation/ironsource/IronSourceAdapterUtils.java
@@ -14,8 +14,23 @@
 
 package com.google.ads.mediation.ironsource;
 
-import android.os.Handler;
-import android.os.Looper;
+import static com.google.ads.mediation.ironsource.IronSourceConstants.TAG;
+import static com.google.ads.mediation.ironsource.IronSourceMediationAdapter.ERROR_DOMAIN;
+import static com.google.ads.mediation.ironsource.IronSourceMediationAdapter.ERROR_INVALID_SERVER_PARAMETERS;
+import static com.google.ads.mediation.ironsource.IronSourceMediationAdapter.ERROR_REQUIRES_ACTIVITY_CONTEXT;
+
+import android.app.Activity;
+import android.content.Context;
+import android.text.TextUtils;
+import android.util.Log;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import com.google.android.gms.ads.AdError;
+import com.google.android.gms.ads.AdSize;
+import com.google.android.gms.ads.MediationUtils;
+import com.ironsource.mediationsdk.ISBannerSize;
+import java.util.ArrayList;
+import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * The {@link IronSourceAdapterUtils} class provides the publisher an ability to pass Activity to
@@ -23,45 +38,58 @@ import android.os.Looper;
  */
 public class IronSourceAdapterUtils {
 
-  /**
-   * Adapter class name for logging.
-   */
-  static final String TAG = IronSourceMediationAdapter.class.getSimpleName();
+  @Nullable
+  public static ISBannerSize getISBannerSizeFromGoogleAdSize(
+      @NonNull Context context, @NonNull AdSize adSize) {
+    ArrayList<AdSize> potentials = new ArrayList<>();
+    potentials.add(AdSize.BANNER);
+    potentials.add(AdSize.MEDIUM_RECTANGLE);
+    potentials.add(AdSize.LARGE_BANNER);
 
-  /**
-   * Key to obtain App Key, required for initializing IronSource SDK.
-   */
-  static final String KEY_APP_KEY = "appKey";
-
-  /**
-   * Key to obtain the IronSource Instance ID, required to show IronSource ads.
-   */
-  static final String KEY_INSTANCE_ID = "instanceId";
-
-  /**
-   * Default IronSource instance ID.
-   */
-  static final String DEFAULT_INSTANCE_ID = "0";
-
-  /**
-   * Constant used for IronSource internal reporting.
-   */
-  static final String MEDIATION_NAME = "AdMob";
-
-  /**
-   * Constant used for IronSource adapter version internal reporting
-   */
-  static final String ADAPTER_VERSION_NAME = "310";
-
-  /**
-   * UI thread handler used to send callbacks with AdMob interface.
-   */
-  private static Handler uiHandler;
-
-  static synchronized void sendEventOnUIThread(Runnable runnable) {
-    if (uiHandler == null) {
-      uiHandler = new Handler(Looper.getMainLooper());
+    AdSize closestSize = MediationUtils.findClosestSize(context, adSize, potentials);
+    if (closestSize == null) {
+      return null;
     }
-    uiHandler.post(runnable);
+
+    if (AdSize.BANNER.equals(closestSize)) {
+      return ISBannerSize.BANNER;
+    } else if (AdSize.MEDIUM_RECTANGLE.equals(closestSize)) {
+      return ISBannerSize.RECTANGLE;
+    } else if (AdSize.LARGE_BANNER.equals(closestSize)) {
+      return ISBannerSize.LARGE;
+    }
+
+    /* If non of the predefined sizes are matched, return a new IronSource size for the closest size
+    returned by Admob. */
+    return new ISBannerSize(closestSize.getWidth(), closestSize.getHeight());
+  }
+
+  public static AdError validateIronSourceAdLoadParams(
+      @NonNull Context context, @NonNull String instanceID) {
+    // Check that context is an Activity.
+    if (!(context instanceof Activity)) {
+      String errorMessage =
+          ERROR_REQUIRES_ACTIVITY_CONTEXT + "IronSource requires an Activity context to load ads.";
+      Log.w(TAG, errorMessage);
+      AdError contextError =
+          new AdError(ERROR_REQUIRES_ACTIVITY_CONTEXT, errorMessage, ERROR_DOMAIN);
+      return contextError;
+    }
+
+    // Check validity of instance ID.
+    if (TextUtils.isEmpty(instanceID)) {
+      AdError loadError =
+          new AdError(
+              ERROR_INVALID_SERVER_PARAMETERS, "Missing or invalid instance ID.", ERROR_DOMAIN);
+      return loadError;
+    }
+
+    return null;
+  }
+
+  public static <T> boolean canLoadIronSourceAdInstance(
+      @NonNull String instanceId, @NonNull ConcurrentHashMap<String, T> instanceMap) {
+    T adUnit = instanceMap.get(instanceId);
+    return (adUnit == null);
   }
 }