+++ ThirdPartyAdapters/ironsource/ironsource/src/main/java/com/google/ads/mediation/ironsource/IronSourceMediationAdapter.java
@@ -14,10 +14,10 @@
 
 package com.google.ads.mediation.ironsource;
 
-import static com.google.ads.mediation.ironsource.IronSourceAdapterUtils.DEFAULT_INSTANCE_ID;
-import static com.google.ads.mediation.ironsource.IronSourceAdapterUtils.KEY_APP_KEY;
-import static com.google.ads.mediation.ironsource.IronSourceAdapterUtils.KEY_INSTANCE_ID;
-import static com.google.ads.mediation.ironsource.IronSourceAdapterUtils.TAG;
+import static com.google.ads.mediation.ironsource.IronSourceConstants.ADAPTER_VERSION_NAME;
+import static com.google.ads.mediation.ironsource.IronSourceConstants.KEY_APP_KEY;
+import static com.google.ads.mediation.ironsource.IronSourceConstants.MEDIATION_NAME;
+import static com.google.ads.mediation.ironsource.IronSourceConstants.TAG;
 
 import android.app.Activity;
 import android.content.Context;
@@ -26,88 +26,68 @@ import android.text.TextUtils;
 import android.util.Log;
 import androidx.annotation.IntDef;
 import androidx.annotation.NonNull;
-import com.google.ads.mediation.ironsource.IronSourceManager.InitializationCallback;
 import com.google.android.gms.ads.AdError;
 import com.google.android.gms.ads.VersionInfo;
 import com.google.android.gms.ads.mediation.Adapter;
 import com.google.android.gms.ads.mediation.InitializationCompleteCallback;
 import com.google.android.gms.ads.mediation.MediationAdLoadCallback;
+import com.google.android.gms.ads.mediation.MediationBannerAd;
+import com.google.android.gms.ads.mediation.MediationBannerAdCallback;
+import com.google.android.gms.ads.mediation.MediationBannerAdConfiguration;
 import com.google.android.gms.ads.mediation.MediationConfiguration;
+import com.google.android.gms.ads.mediation.MediationInterstitialAd;
+import com.google.android.gms.ads.mediation.MediationInterstitialAdCallback;
+import com.google.android.gms.ads.mediation.MediationInterstitialAdConfiguration;
 import com.google.android.gms.ads.mediation.MediationRewardedAd;
 import com.google.android.gms.ads.mediation.MediationRewardedAdCallback;
 import com.google.android.gms.ads.mediation.MediationRewardedAdConfiguration;
-import com.google.android.gms.ads.rewarded.RewardItem;
-import com.ironsource.mediationsdk.logger.IronSourceError;
+import com.ironsource.mediationsdk.IronSource;
 import com.ironsource.mediationsdk.utils.IronSourceUtils;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.util.HashSet;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicBoolean;
 
-public class IronSourceMediationAdapter extends Adapter
-    implements MediationRewardedAd, IronSourceAdapterListener {
+public class IronSourceMediationAdapter extends Adapter {
+
+  private static final AtomicBoolean isInitialized = new AtomicBoolean(false);
 
   // region Error codes
-  // IronSource adapter error domain.
+  /** IronSource adapter error domain. */
   public static final String ERROR_DOMAIN = "com.google.ads.mediation.ironsource";
 
-  // IronSource SDK error domain.
+  /** IronSource SDK error domain. */
   public static final String IRONSOURCE_SDK_ERROR_DOMAIN = "com.ironsource.mediationsdk";
 
   @Retention(RetentionPolicy.SOURCE)
   @IntDef(
       value = {
-          ERROR_INVALID_SERVER_PARAMETERS,
-          ERROR_REQUIRES_ACTIVITY_CONTEXT,
-          ERROR_AD_ALREADY_LOADED,
-          ERROR_AD_SHOW_UNAUTHORIZED,
+        ERROR_INVALID_SERVER_PARAMETERS,
+        ERROR_REQUIRES_ACTIVITY_CONTEXT,
+        ERROR_AD_ALREADY_LOADED,
+        ERROR_BANNER_SIZE_MISMATCH,
+        ERROR_SDK_NOT_INITIALIZED
       })
-  public @interface AdapterError {
+  public @interface AdapterError {}
 
-  }
-
-  /**
-   * Server parameters (e.g. placement ID) are nil.
-   */
+  /** Server parameters (e.g. instance ID) are nil. */
   public static final int ERROR_INVALID_SERVER_PARAMETERS = 101;
 
-  /**
-   * IronSource requires an {@link Activity} context to initialize their SDK.
-   */
+  /** IronSource requires an {@link Activity} context to initialize their SDK. */
   public static final int ERROR_REQUIRES_ACTIVITY_CONTEXT = 102;
 
-  /**
-   * IronSource can only load 1 ad per IronSource instance ID.
-   */
+  /** IronSource can only load 1 ad per IronSource instance ID. */
   public static final int ERROR_AD_ALREADY_LOADED = 103;
 
-  /**
-   * IronSource adapter does not have authority to show an ad instance.
-   */
-  public static final int ERROR_AD_SHOW_UNAUTHORIZED = 104;
+  /** Banner size mismatch. */
+  public static final int ERROR_BANNER_SIZE_MISMATCH = 105;
+
+  /** IronSource SDK isn't initialized. */
+  public static final int ERROR_SDK_NOT_INITIALIZED = 106;
+
   // endregion
 
-  /**
-   * Mediation listener used to forward rewarded ad events from IronSource SDK to Google Mobile Ads
-   * SDK while ad is presented
-   */
-  private MediationRewardedAdCallback mediationRewardedAdCallback;
-
-  /**
-   * Mediation listener used to forward rewarded ad events from IronSource SDK to Google Mobile Ads
-   * SDK for loading phases of the ad
-   */
-  private MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback>
-      mediationAdLoadCallback;
-
-  /**
-   * This is the id of the rewarded video instance requested.
-   */
-  private String instanceID;
-
-  /**
-   * MediationRewardedAd implementation.
-   */
   @NonNull
   @Override
   public VersionInfo getSDKVersionInfo() {
@@ -158,9 +138,15 @@ public class IronSourceMediationAdapter extends Adapter
   }
 
   @Override
-  public void initialize(@NonNull Context context,
+  public void initialize(
+      @NonNull Context context,
       @NonNull final InitializationCompleteCallback initializationCompleteCallback,
       @NonNull List<MediationConfiguration> mediationConfigurations) {
+
+    if (isInitialized.get()) {
+      initializationCompleteCallback.onInitializationSucceeded();
+      return;
+    }
 
     HashSet<String> appKeys = new HashSet<>();
     for (MediationConfiguration configuration : mediationConfigurations) {
@@ -174,248 +160,122 @@ public class IronSourceMediationAdapter extends Adapter
 
     int count = appKeys.size();
     if (count <= 0) {
-      AdError initializationError = new AdError(ERROR_INVALID_SERVER_PARAMETERS,
-          "Missing or invalid app key.", ERROR_DOMAIN);
+      AdError initializationError =
+          new AdError(ERROR_INVALID_SERVER_PARAMETERS, "Missing or invalid app key.", ERROR_DOMAIN);
       initializationCompleteCallback.onInitializationFailed(initializationError.getMessage());
       return;
     }
 
     // Having multiple app keys is not considered an error.
     String appKey = appKeys.iterator().next();
+
+    if (TextUtils.isEmpty(appKey)) {
+      AdError initializationError =
+          new AdError(ERROR_INVALID_SERVER_PARAMETERS, "Missing or invalid app key.", ERROR_DOMAIN);
+      initializationCompleteCallback.onInitializationFailed(initializationError.getMessage());
+      return;
+    }
+
     if (count > 1) {
-      String message = String
-          .format("Multiple '%s' entries found: %s. Using '%s' to initialize the IronSource SDK.",
+      String message =
+          String.format(
+              "Multiple '%s' entries found: %s. Using app key '%s' to initialize the IronSource SDK.",
               KEY_APP_KEY, appKeys, appKey);
       Log.w(TAG, message);
     }
 
-    IronSourceManager.getInstance().initIronSourceSDK(context, appKey,
-        new IronSourceManager.InitializationCallback() {
-          @Override
-          public void onInitializeSuccess() {
-            initializationCompleteCallback.onInitializationSucceeded();
-          }
+    IronSource.setMediationType(MEDIATION_NAME + ADAPTER_VERSION_NAME);
+    Log.d(TAG, "Initializing IronSource SDK with app key: " + appKey);
+    IronSource.initISDemandOnly(
+        context,
+        appKey,
+        IronSource.AD_UNIT.INTERSTITIAL,
+        IronSource.AD_UNIT.REWARDED_VIDEO,
+        IronSource.AD_UNIT.BANNER);
+    isInitialized.set(true);
+    initializationCompleteCallback.onInitializationSucceeded();
 
-          @Override
-          public void onInitializeError(@NonNull AdError initializationError) {
-            initializationCompleteCallback.onInitializationFailed(initializationError.getMessage());
-          }
-        });
+    IronSource.setISDemandOnlyInterstitialListener(
+        IronSourceInterstitialAd.getIronSourceInterstitialListener());
+    IronSource.setISDemandOnlyRewardedVideoListener(
+        IronSourceRewardedAd.getIronSourceRewardedListener());
   }
 
   @Override
   public void loadRewardedAd(
       @NonNull MediationRewardedAdConfiguration mediationRewardedAdConfiguration,
-      @NonNull final MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback>
-          mediationAdLoadCallback) {
+      @NonNull
+          final MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback>
+              mediationAdLoadCallback) {
+    if (!isInitialized.get()) {
+      AdError adError =
+          new AdError(
+              ERROR_SDK_NOT_INITIALIZED,
+              "Cannot load rewarded ad when the IronSource SDK is not initialized.",
+              ERROR_DOMAIN);
+      mediationAdLoadCallback.onFailure(adError);
+      return;
+    }
 
-    Bundle serverParameters = mediationRewardedAdConfiguration.getServerParameters();
-    Context context = mediationRewardedAdConfiguration.getContext();
-    String appKey = serverParameters.getString(KEY_APP_KEY);
-    this.instanceID = serverParameters.getString(KEY_INSTANCE_ID, DEFAULT_INSTANCE_ID);
-
-    IronSourceManager.getInstance().initIronSourceSDK(context, appKey,
-        new InitializationCallback() {
-          @Override
-          public void onInitializeSuccess() {
-            IronSourceMediationAdapter.this.mediationAdLoadCallback = mediationAdLoadCallback;
-            Log.d(TAG,
-                String.format("Loading IronSource rewarded ad with instance ID: %s", instanceID));
-            IronSourceManager.getInstance()
-                .loadRewardedVideo(context, instanceID, IronSourceMediationAdapter.this);
-          }
-
-          @Override
-          public void onInitializeError(@NonNull AdError initializationError) {
-            Log.e(TAG, initializationError.getMessage());
-            mediationAdLoadCallback.onFailure(initializationError);
-          }
-        });
+    IronSourceRewardedAd ironSourceRewardedAd =
+        new IronSourceRewardedAd(mediationRewardedAdConfiguration, mediationAdLoadCallback);
+    ironSourceRewardedAd.loadAd();
   }
 
   @Override
   public void loadRewardedInterstitialAd(
       @NonNull MediationRewardedAdConfiguration mediationRewardedAdConfiguration,
-      @NonNull MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback>
-          mediationAdLoadCallback) {
+      @NonNull
+          MediationAdLoadCallback<MediationRewardedAd, MediationRewardedAdCallback>
+              mediationAdLoadCallback) {
     // IronSource Rewarded Interstitial ads use the same Rewarded Video API.
-    Log.d(TAG, "IronSource adapter was asked to load a rewarded interstitial ad. "
-        + "Using the rewarded ad request flow to load the ad to attempt to load a "
-        + "rewarded interstitial ad from IronSource.");
+    Log.d(
+        TAG,
+        "IronSource adapter was asked to load a rewarded interstitial ad. "
+            + "Using the rewarded ad request flow to load the ad to attempt to load a "
+            + "rewarded interstitial ad from IronSource.");
     loadRewardedAd(mediationRewardedAdConfiguration, mediationAdLoadCallback);
   }
 
   @Override
-  public void showAd(@NonNull Context context) {
-    Log.d(TAG,
-        String.format("Showing IronSource rewarded ad for instance ID: %s", this.instanceID));
-    IronSourceManager.getInstance()
-        .showRewardedVideo(this.instanceID, IronSourceMediationAdapter.this);
-  }
+  public void loadInterstitialAd(
+      @NonNull MediationInterstitialAdConfiguration mediationInterstitialAdConfiguration,
+      @NonNull
+          MediationAdLoadCallback<MediationInterstitialAd, MediationInterstitialAdCallback>
+              mediationAdLoadCallback) {
+    if (!isInitialized.get()) {
+      AdError loadError =
+          new AdError(
+              ERROR_SDK_NOT_INITIALIZED,
+              "Cannot load interstitial ad when the IronSource SDK is not initialized.",
+              IRONSOURCE_SDK_ERROR_DOMAIN);
+      mediationAdLoadCallback.onFailure(loadError);
+      return;
+    }
 
-  // region ISDemandOnlyRewardedVideoListener implementation.
-  public void onRewardedVideoAdLoadSuccess(String instanceId) {
-    Log.d(TAG, String.format("IronSource rewarded ad loaded for instance ID: %s", instanceId));
-
-    IronSourceAdapterUtils.sendEventOnUIThread(
-        new Runnable() {
-          @Override
-          public void run() {
-            if (mediationAdLoadCallback != null) {
-              mediationRewardedAdCallback =
-                  mediationAdLoadCallback.onSuccess(IronSourceMediationAdapter.this);
-            }
-          }
-        });
-  }
-
-  public void onRewardedVideoAdLoadFailed(String instanceId, IronSourceError ironSourceError) {
-    AdError loadError = new AdError(ironSourceError.getErrorCode(),
-        ironSourceError.getErrorMessage(), IRONSOURCE_SDK_ERROR_DOMAIN);
-    String errorMessage = String
-        .format("IronSource failed to load rewarded ad for instance ID: %s. Error: %s", instanceId,
-            loadError.getMessage());
-    Log.e(TAG, errorMessage);
-
-    IronSourceAdapterUtils.sendEventOnUIThread(
-        new Runnable() {
-          @Override
-          public void run() {
-            if (mediationAdLoadCallback != null) {
-              mediationAdLoadCallback.onFailure(loadError);
-            }
-          }
-        });
-  }
-
-  public void onRewardedVideoAdOpened(final String instanceId) {
-    Log.d(TAG, String.format("IronSource rewarded ad opened for instance ID: %s", instanceId));
-
-    IronSourceAdapterUtils.sendEventOnUIThread(
-        new Runnable() {
-          @Override
-          public void run() {
-            if (mediationRewardedAdCallback != null) {
-              mediationRewardedAdCallback.onAdOpened();
-              mediationRewardedAdCallback.onVideoStart();
-              mediationRewardedAdCallback.reportAdImpression();
-            }
-          }
-        });
-  }
-
-  public void onRewardedVideoAdClosed(String instanceId) {
-    Log.d(TAG, String.format("IronSource rewarded ad closed for instance ID: %s", instanceId));
-
-    IronSourceAdapterUtils.sendEventOnUIThread(
-        new Runnable() {
-          @Override
-          public void run() {
-            if (mediationRewardedAdCallback != null) {
-              mediationRewardedAdCallback.onAdClosed();
-            }
-          }
-        });
-  }
-
-  public void onRewardedVideoAdRewarded(String instanceId) {
-    final IronSourceReward reward = new IronSourceReward();
-    Log.d(
-        TAG,
-        String.format(
-            "IronSource rewarded ad received reward: %d %s, for instance ID: %s",
-            reward.getAmount(), reward.getType(), instanceId));
-
-    IronSourceAdapterUtils.sendEventOnUIThread(
-        new Runnable() {
-          @Override
-          public void run() {
-            if (mediationRewardedAdCallback != null) {
-              mediationRewardedAdCallback.onVideoComplete();
-              mediationRewardedAdCallback.onUserEarnedReward(reward);
-            }
-          }
-        });
-  }
-
-  public void onRewardedVideoAdShowFailed(String instanceId, IronSourceError ironSourceError) {
-    AdError showError = new AdError(ironSourceError.getErrorCode(),
-        ironSourceError.getErrorMessage(), IRONSOURCE_SDK_ERROR_DOMAIN);
-    String errorMessage = String
-        .format("IronSource failed to show rewarded ad for instance ID: %s. Error: %s", instanceId,
-            showError.getMessage());
-    Log.e(TAG, errorMessage);
-
-    IronSourceAdapterUtils.sendEventOnUIThread(
-        new Runnable() {
-          @Override
-          public void run() {
-            if (mediationRewardedAdCallback != null) {
-              mediationRewardedAdCallback.onAdFailedToShow(showError);
-            }
-          }
-        });
-  }
-
-  public void onRewardedVideoAdClicked(String instanceId) {
-    Log.d(TAG, String.format("IronSource rewarded ad clicked for instance ID: %s", instanceId));
-
-    IronSourceAdapterUtils.sendEventOnUIThread(
-        new Runnable() {
-          @Override
-          public void run() {
-            if (mediationRewardedAdCallback != null) {
-              mediationRewardedAdCallback.reportAdClicked();
-            }
-          }
-        });
-  }
-  // endregion
-
-  // region IronSourceAdapterListener implementation.
-  @Override
-  public void onAdFailedToLoad(@NonNull AdError loadError) {
-    Log.e(TAG, loadError.getMessage());
-    IronSourceAdapterUtils.sendEventOnUIThread(
-        new Runnable() {
-          @Override
-          public void run() {
-            if (mediationAdLoadCallback != null) {
-              mediationAdLoadCallback.onFailure(loadError);
-            }
-          }
-        });
+    IronSourceInterstitialAd ironSourceInterstitialAd =
+        new IronSourceInterstitialAd(mediationInterstitialAdConfiguration, mediationAdLoadCallback);
+    ironSourceInterstitialAd.loadAd();
   }
 
   @Override
-  public void onAdFailedToShow(@NonNull AdError showError) {
-    Log.e(TAG, showError.getMessage());
-    IronSourceAdapterUtils.sendEventOnUIThread(
-        new Runnable() {
-          @Override
-          public void run() {
-            if (mediationRewardedAdCallback != null) {
-              mediationRewardedAdCallback.onAdFailedToShow(showError);
-            }
-          }
-        });
-  }
-  // endregion
-
-  /**
-   * A {@link RewardItem} used to map IronSource reward to Google's reward.
-   */
-  static class IronSourceReward implements RewardItem {
-
-    @NonNull
-    @Override
-    public String getType() {
-      return "";
+  public void loadBannerAd(
+      @NonNull MediationBannerAdConfiguration mediationBannerAdConfiguration,
+      @NonNull
+          MediationAdLoadCallback<MediationBannerAd, MediationBannerAdCallback>
+              mediationAdLoadCallback) {
+    if (!isInitialized.get()) {
+      AdError loadError =
+          new AdError(
+              ERROR_SDK_NOT_INITIALIZED,
+              "Cannot load banner ad when the IronSource SDK is not initialized.",
+              IRONSOURCE_SDK_ERROR_DOMAIN);
+      mediationAdLoadCallback.onFailure(loadError);
+      return;
     }
 
-    @Override
-    public int getAmount() {
-      return 1;
-    }
+    IronSourceBannerAd ironSourceBannerAd =
+        new IronSourceBannerAd(mediationBannerAdConfiguration, mediationAdLoadCallback);
+    ironSourceBannerAd.loadAd();
   }
 }