+++ ThirdPartyAdapters/ironsource/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## IronSource Android Mediation Adapter Changelog
 
-#### Version 7.4.0.0
+#### Next version
+- Added waterfall support for banner (includes MREC) ad format.
+- Updated the adapter to use the new mediation APIs.
+
+#### Version 7.4.0.0 (In progress)
 - Verified compatibility with ironSource SDK version 7.4.0.
 - Updated the minimum required Google Mobile Ads SDK version to 22.2.0.